pyproject.toml
[tool.poetry]
name = "hpi-ii-project-2022"
version = "0.1.0"
description = ""
authors = ["raminqaf <ramin.gharib@bakdata.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bakdata/hpi-ii-project-2022"
packages = [
    { include = "rb_crawler" }
]

[tool.poetry.dependencies]
python = "^3.8"
async_generator = "1.10"
attrs = "21.4.0"
Automat = "20.2.0"
beautifulsoup4 = "4.11.1"
cachetools = "5.0.0"
certifi = "2021.10.8"
cffi = "1.15.0"
charset-normalizer = "2.0.12"
chromedriver-binary = "99.0.4844.51.0"
confluent-kafka = "1.8.2"
constantly = "15.1.0"
cryptography = "36.0.2"
cssselect = "1.1.0"
filelock = "3.6.0"
google = "3.0.0"
google-api-core = "2.7.1"
google-api-python-client = "2.43.0"
google-auth = "2.6.3"
google-auth-httplib2 = "0.1.0"
googleapis-common-protos = "1.56.0"
h11 = "0.13.0"
httplib2 = "0.20.4"
hyperlink = "21.0.0"
idna = "3.3"
incremental = "21.3.0"
itemadapter = "0.5.0"
itemloaders = "1.0.4"
jmespath = "1.0.0"
lxml = "4.8.0"
outcome = "1.1.0"
parsel = "1.6.0"
Protego = "0.2.1"
protobuf = "3.20.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
PyDispatcher = "2.0.5"
pyOpenSSL = "22.0.0"
pyparsing = "3.0.8"
PySocks = "1.7.1"
queuelib = "1.6.2"
requests = "2.27.1"
requests-file = "1.5.1"
rsa = "4.8"
Scrapy = "2.6.1"
service-identity = "21.1.0"
six = "1.16.0"
sniffio = "1.2.0"
sortedcontainers = "2.4.0"
soupsieve = "2.3.2"
tldextract = "3.2.0"
trio = "0.20.0"
trio-websocket = "0.9.2"
Twisted = "22.2.0"
typing-extensions = "4.1.1"
uritemplate = "4.1.1"
urllib3 = "1.26.9"
w3lib = "1.22.0"
wsproto = "1.1.0"
"zope.interface" = "5.4.0"
click = "8.0.2"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true, python = "^3.8" }
isort = "^5.4"
flake8 = "^3.8"
autoflake = "^1.4"

[tool.isort]
profile = "black"
force_single_line = true
atomic = true
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
src_paths = ["rb_crawler"]
skip_glob = ['build/*']
filter_files = true
known_first_party = "rb_crawler"
line_length = 100

[tool.black]
line-length = 120
target-version = ["py38", "py39"]
exclude = '''
/(
   build
 | .eggs
 | .git
 | .pytest_cache
 | dist
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
